OpenCV Load & Show Image
Fragment of code that actually works.  It's written in Python.  To run it, first download it and store it with a .py extension in a directory that you like, then open your terminal, move to the directory that you used to store the file using cd.  For example if you store it in your desktop, type: 
>cd Desktop/
Then type:
>python fileName.py --image Path the image/imageName.png